Add Repentogon check, move version checking to its own file
