+++ BasementRenovator.py
@@ -139,33 +139,11 @@ from src.lookup import EntityLookup, MainLookup
 import src.anm2 as anm2
 from src.constants import *
 from src.util import *
+from src.version import *
 
 ########################
 #       XML Data       #
 ########################
-
-
-def getGameVersion():
-    """
-    Returns the current compatibility mode, and the sub version if it exists
-    """
-    # default mode if not set
-    mode = settings.value("CompatibilityMode", "Repentance+")
-
-    return mode
-
-
-def willLaunchREPENTOGON():
-    exePath: str | None = settings.value("CustomExePath")
-    return exePath and exePath.lower().endswith("repentogonlauncher.exe")
-
-
-def canUseREPENTOGON():
-    if getGameVersion() != "Repentance+":
-        return False
-
-    return willLaunchREPENTOGON()
-
 
 STEAM_PATH = None
